Fixed setting and getting subject IDs
Suject IDs are recorded in freadInterval (queryBTNode.c) as a bit vector
which was simplified to hold at most WORDSIZE IDs, which was 32 in this
case. This didn't work in general, not even for the 42 HIV genomes of
the test dataset. This is now fixed, similarly in
inorder (queryBTNode.c), where the subject names are looked up again.