Fixed bug in union
We swapped slices a and b to ensure |a| < |b| in order to construct the
dictionary from the smaller of the two slices. As a result, we might
swap their addresses, which led to spurious annotations. Now we still
construct the dictionary from the smaller of a and b, but the union
returned is a fresh slice rather than the address of either a or b.